// src/types/payment.d.ts
export const VALID_PAYMENT_FILTER_KEYS = [
  'originatorName',
  'originatorAccount',
  'originatorBankRoutingCode',
  'recipientName',
  'recipientAccount',
  'recipientBankRoutingCode',
  'reference',
  'traceNumber',
  'externalId',
  'clientId',
  'dateFormat',
  'locale',
  'originatedBy',
  'paymentRail',
  'paymentType',
  'fromValueDate',
  'toValueDate',
  'fromExecuteDate',
  'toExecuteDate',
  'status',
  'fromReturnDate',
  'toReturnDate',
  'isSettlement',
  'orderBy',
  'sortOrder'
] as const;